+++ GenericServices/StatusGenericHandler.cs
@@ -110,12 +110,13 @@ namespace GenericServices
         /// This is a simple method to output all the errors as a single string - null if no errors
         /// Useful for feeding back all the errors in a single exception (also nice in unit testing)
         /// </summary>
-        /// <param name="seperator"></param>
-        /// <returns>a single string with all errors seperated by the 'seperator' string</returns>
-        public string GetAllErrors(string seperator = "\n")
+        /// <param name="separator"></param>
+        /// <returns>a single string with all errors separated by the 'separator' string</returns>
+        public string GetAllErrors(string separator = null)
         {
+            separator = separator ?? Environment.NewLine;
             return _errors.Any() 
-                ? string.Join(seperator, Errors) 
+                ? string.Join(separator, Errors) 
                 : null;
         }
     }